netpol: Fix positions of iptables rules
This change ensures that all:

* service IP rules
* TCP node ports
* UDP node ports
* external IPs

are submitted after each other, regardless of IP family.

Before this change, with hardcoded positions, secondary service IP was
replacing the primary one (sec. service IP had position 1, so the
primary was moved to position 2), then the first TCP node port was also
replacing it (TCP node port had position 2, so it moved down the primary
service IP to position 3). That made a disorder in iptables rules.

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>